+++ src/array/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+import equals from './equals'
+
+export {
+    equals,
+}