+++ src/index.js
@@ -2,9 +2,11 @@
 
 import clone from './core/clone'
 import toString from './core/toString'
+import equals from './core/equals'
 
 
 export {
     clone,
     toString,
+    equals,
 }