+++ src/object/equals.js
@@ -1,55 +1,123 @@
 'use strict'
 
+import {isNumeric} from '../is'
 
 /**
- * @method
- * Shallow compares the contents of 2 objects using strict equality. Objects are
- * considered equal if they both have the same set of properties and the
- * value for those properties equals the other in the corresponding object.
+ * Compares the contents of 2 or more objects using strict equality.
  *
- *     // Returns true
- *     equals({
- *         foo: 1,
- *         bar: 2
- *     }, {
- *         foo: 1,
- *         bar: 2
- *     });
- *
- * @param {Object} object1
- * @param {Object} object2
- * @return {Boolean} `true` if the objects are equal.
- * @todo write tests
+ * @param {Object} object_1
+ * @param {Object...} object_N
+ * @returns {boolean}
  */
-export default (function () {
-    const check = function (o1, o2) {
-        let key
+export default function equal() {
+    let i, l, leftChain, rightChain
 
-        for (key in o1) {
-            if (Object.prototype.hasOwnProperty.call(o1, key)) {
-                if (o1[key] !== o2[key]) {
+    function compare2Objects(x, y) {
+        let p
+
+        // remember that NaN === NaN returns false
+        // and isNaN(undefined) returns true
+        isNumeric()
+        if (isNaN(x) && isNaN(y) && typeof x === 'number' && typeof y === 'number') {
+            return true
+        }
+
+        // Compare primitives and functions.
+        // Check if both arguments link to the same object.
+        // Especially useful on the step where we compare prototypes
+        if (x === y) {
+            return true
+        }
+
+        // Works in case when functions are created in constructor.
+        // Comparing dates is a common scenario. Another built-ins?
+        // We can even handle functions passed across iframes
+        if ((typeof x === 'function' && typeof y === 'function') ||
+            (x instanceof Date && y instanceof Date) ||
+            (x instanceof RegExp && y instanceof RegExp) ||
+            (x instanceof String && y instanceof String) ||
+            (x instanceof Number && y instanceof Number)) {
+            return x.toString() === y.toString()
+        }
+
+        // At last checking prototypes as good as we can
+        if (!(x instanceof Object && y instanceof Object)) {
+            return false
+        }
+
+        if (Object.prototype.isPrototypeOf.call(x, y) || Object.prototype.isPrototypeOf.call(y, x)) {
+            return false
+        }
+
+        if (x.constructor !== y.constructor) {
+            return false
+        }
+
+        if (x.prototype !== y.prototype) {
+            return false
+        }
+
+        // Check for infinitive linking loops
+        if (leftChain.indexOf(x) > -1 || rightChain.indexOf(y) > -1) {
+            return false
+        }
+
+        // Quick checking of one object being a subset of another.
+        // todo: cache the structure of arguments[0] for performance
+        for (p in y) {
+            if (Object.prototype.isPrototypeOf.call(y, p) !== Object.prototype.isPrototypeOf.call(x, p)) {
+                return false
+            } else if (typeof y[p] !== typeof x[p]) {
+                return false
+            }
+        }
+
+        for (p in x) {
+            if (Object.prototype.isPrototypeOf.call(y, p) !== Object.prototype.isPrototypeOf.call(x, p)) {
+                return false
+            } else if (typeof y[p] !== typeof x[p]) {
+                return false
+            }
+
+            switch (typeof (x[p])) {
+            case 'object':
+            case 'function':
+
+                leftChain.push(x)
+                rightChain.push(y)
+
+                if (!compare2Objects(x[p], y[p])) {
                     return false
                 }
+
+                leftChain.pop()
+                rightChain.pop()
+                break
+
+            default:
+                if (x[p] !== y[p]) {
+                    return false
+                }
+                break
             }
         }
 
         return true
     }
 
-    return function (object1, object2) {
-        // Short circuit if the same object is passed twice
-        if (object1 === object2) {
-            return true
-        }
+    if (arguments.length <= 1) {
+        throw  Error('Need two or more arguments to compare')
+    }
 
-        if (object1 && object2) {
-            // Do the second check because we could have extra keys in
-            // object2 that don't exist in object1.
-            return check(object1, object2) && check(object2, object1)
-        } else if (!object1 && !object2) {
-            return object1 === object2
-        } else {
+    for (i = 1, l = arguments.length; i < l; i++) {
+
+        leftChain = [] // @Todo: this can be cached
+        rightChain = []
+
+        if (!compare2Objects(arguments[0], arguments[i])) {
             return false
         }
     }
-})()
+
+    return true
+}